first steps towards user page
