--- src/routes/friends/+page.svelte
@@ -1,1 +1,0 @@
-Hi, you have no friends.
